refactor(graphql): Support MoveType built from MoveTypeLayout
## Description

The output from extracting one MoveValue from another is an
`OwnedSlice` -- a vector of bytes accompanied by its layout. Previously
we discarded the layout information, collapsing it into just a type when
we wrapped this in a `MoveType`, which meant that if you needed to e.g.
output the extracted value as JSON, we needed to go to the package
resolver again to fetch its layout.

This change updates the representation of `MoveType` to allow it to hold
on to a full layout if it was initialised with one. This means that if
its layout is requested again, it can simply clone the existing layout
rather than re-resolve it.

## Test plan

Existing tests.