display: entrypoint for dynamic field names
## Description

Adding a new entrypoint for parsing a literal that can be interpreted as
a dynamic field name (deriving an ObjectID from an expression that can
be interpreted into something that has a type and can be serialized into
BCS).

This will be used to support accepting dynamic fields as strings, rather
than requiring they be BCS serialized beforehand.

## Test plan

New unit tests:

```
$ cargo nextest run -p sui-display
```